requirements.txt
falcon>=4.0,<5
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
gunicorn>=22.0,<24
pytest>=8,<10
jsonschema>=4.20,<5
